+++ oxlint.config.ts
@@ -1,0 +1,14 @@
+import { defineConfig } from "oxlint"
+
+import config from "./src/index.ts"
+
+export default defineConfig({
+	"extends": [config],
+	"env": {
+		node: true,
+	},
+	"ignorePatterns": [
+		`dist`,
+		`fixtures`,
+	],
+})