+++ src/plugins/simple-import-sort.ts
@@ -1,0 +1,1 @@
+export { default } from "eslint-plugin-simple-import-sort"