+++ vitest.config.ts
@@ -2,7 +2,10 @@ import { defineConfig } from "vitest/config"
 
 export default defineConfig({
 	test: {
-		include: [`lib/**/*.test.js`],
+		include: [
+			`src/**/*.test.ts`,
+			`tests/**/*.test.ts`,
+		],
 		watch: false,
 	},
 })